requirements.txt
# pin ranges so pip doesnt pull random breaking stuff
langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-openai>=0.2.0,<0.4
langgraph>=0.2.0,<0.4
rouge-score>=0.1.2
mlflow>=2.14.0,<3
python-dotenv>=1.0.0
